online shopping system ood fix

+++ Leetcode Practices/object oriented design/grok object oriented design interview/4. Source Codes/Amazon - Online Shopping System/model/Item.java
@@ -3,5 +3,5 @@ public class Item {
     private int quantity;
     private double price;
 
-    public boolean updateQuantity(long quantity);
+    public boolean updateQuantity(int quantity);
 }
