+++ Leetcode Practices/object oriented design/grok object oriented design interview/4. Source Codes/Amazon - Online Shopping System/model/ShoppingCart.java
@@ -3,7 +3,7 @@ public class ShoppingCart {
 
     public boolean addItem(Item item);
     public boolean removeItem(Item item);
-    public boolean updateItemQuantity(Item item, long quantity);
+    public boolean updateItemQuantity(Item item, int quantity);
     public List<Item> getItems();
     public boolean checkout();
 }
